14. zamiana klasy Node na prywatną, przeniesienie jej do klasy IntLinkedList oraz usuniecie pliku Node.java
